dom/workers/test/test_navigator_workers_hardwareConcurrency.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Navigator.hardwareConcurrency</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  SimpleTest.waitForExplicitFinish();
  var script = "postMessage(navigator.hardwareConcurrency)";
  var url = URL.createObjectURL(new Blob([script]));
  var w = new Worker(url);
  w.onmessage = function(e) {
    var x = e.data;
    is(typeof x, "number", "hardwareConcurrency should be a number.");
    ok(x > 0, "hardwareConcurrency should be greater than 0.");
    SimpleTest.finish();
  }
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
